// client/src/store/actionTypes.ts
export enum ActionTypes {
  GET_BOOKS = "get books",
  GET_BOOKS_SUCCESS = "get books success",
  CREATE_NEW_BOOK = "create new book",
  UPDATE_BOOK = "update book",
  UPDATE_BOOK_SUCCESS = "update book success",
  DELETE_BOOK = "delete book",
  DELETE_BOOK_SUCCESS = "delete book success",
  GET_AUTHORS = "get authors",
  GET_AUTHORS_SUCCESS = "get authors success",
  CREATE_NEW_AUTHOR = "create new author",
  UPDATE_AUTHOR = "update author",
  UPDATE_AUTHOR_SUCCESS = "update author success",
  DELETE_AUTHOR = "delete author",
  DELETE_AUTHOR_SUCCESS = "delete author success",
}